+++ packages/core/client/src/schema-component/antd/markdown/util.ts
@@ -8,8 +8,6 @@
  */
 
 import { useEffect, useState } from 'react';
-import Handlebars from 'handlebars';
-import { replaceVariableValue } from '../../../block-provider/hooks';
 
 export async function parseMarkdown(text: string) {
   if (!text) {
@@ -41,40 +39,4 @@ export function convertToText(markdownText: string) {
   const text = temp.innerText;
   temp = null;
   return text?.replace(/[\n\r]/g, '') || '';
-}
-
-Handlebars.registerHelper('eq', function (arg1, arg2) {
-  return arg1 == arg2;
-});
-Handlebars.registerHelper('toUpperCase', function (str) {
-  return str.toUpperCase();
-});
-
-const getVariablesData = (localVariables) => {
-  const data = {};
-  localVariables.map((v) => {
-    data[v.name] = v.ctx;
-  });
-  return data;
-};
-
-export async function getRenderContent(templateEngine, content, variables, localVariables) {
-  if (templateEngine === 'handlebars') {
-    const renderedContent = Handlebars.compile(content);
-    // 处理渲染后的内容
-    try {
-      const data = getVariablesData(localVariables);
-      return renderedContent({ ...variables.ctxRef.current, ...data });
-    } catch (error) {
-      console.log(error);
-      return content;
-    }
-  } else {
-    try {
-      const html = await replaceVariableValue(content, variables, localVariables);
-      return await parseMarkdown(html);
-    } catch (error) {
-      return content;
-    }
-  }
 }
