+++ packages/core/client/src/schema-component/common/index.ts
@@ -10,3 +10,4 @@
 export * from './dnd-context';
 export * from './sortable-item';
 export * from './show-form-data';
+export { getRenderContent } from './utils/uitls';
