+++ packages/core/client/src/schema-component/common/utils/uitls.tsx
@@ -8,10 +8,12 @@
  */
 
 import _, { every, findIndex, some } from 'lodash';
+import Handlebars from 'handlebars';
 import { VariableOption, VariablesContextType } from '../../../variables/types';
 import { isVariable } from '../../../variables/utils/isVariable';
 import { transformVariableValue } from '../../../variables/utils/transformVariableValue';
 import { getJsonLogic } from '../../common/utils/logic';
+import { replaceVariableValue } from '../../../block-provider/hooks';
 
 type VariablesCtx = {
   /** 当前登录的用户 */
@@ -137,4 +139,40 @@ export const conditionAnalyses = async ({
 export function targetFieldToVariableString(targetField: string[]) {
   // Action 中的联动规则虽然没有 form 上下文但是在这里也使用的是 `$nForm` 变量，这样实现更简单
   return `{{ $nForm.${targetField.join('.')} }}`;
+}
+
+Handlebars.registerHelper('eq', function (arg1, arg2) {
+  return arg1 == arg2;
+});
+Handlebars.registerHelper('toUpperCase', function (str) {
+  return str.toUpperCase();
+});
+
+const getVariablesData = (localVariables) => {
+  const data = {};
+  localVariables.map((v) => {
+    data[v.name] = v.ctx;
+  });
+  return data;
+};
+
+export async function getRenderContent(templateEngine, content, variables, localVariables, defaultParse) {
+  if (content && templateEngine === 'handlebars') {
+    const renderedContent = Handlebars.compile(content);
+    // 处理渲染后的内容
+    try {
+      const data = getVariablesData(localVariables);
+      return renderedContent({ ...variables.ctxRef.current, ...data });
+    } catch (error) {
+      console.log(error);
+      return content;
+    }
+  } else {
+    try {
+      const html = await replaceVariableValue(content, variables, localVariables);
+      return await defaultParse(html);
+    } catch (error) {
+      return content;
+    }
+  }
 }
