+++ packages/plugins/@nocobase/plugin-block-iframe/src/client/Iframe.tsx
@@ -15,6 +15,7 @@ import {
   useParseURLAndParams,
   useRequest,
   useVariables,
+  getRenderContent,
 } from '@nocobase/client';
 import { Card, Spin } from 'antd';
 import React, { useEffect, useState } from 'react';
@@ -32,8 +33,8 @@ function isNumeric(str: string | undefined) {
 }
 
 export const Iframe: any = observer(
-  (props: IIframe & { html?: string; htmlId?: number; mode: string; params?: any }) => {
-    const { url, htmlId, mode = 'url', height, html, params, ...others } = props;
+  (props: IIframe & { html?: string; htmlId?: number; mode: string; params?: any; engine?: string }) => {
+    const { url, htmlId, mode = 'url', height, html, params, engine, ...others } = props;
     const field = useField();
     const { t } = useTranslation();
     const targetHeight = useBlockHeight() || height;
@@ -50,13 +51,15 @@ export const Iframe: any = observer(
     );
     const { parseURLAndParams } = useParseURLAndParams();
     const [src, setSrc] = useState(null);
-
     useEffect(() => {
       const generateSrc = async () => {
         if (mode === 'html') {
-          const targetHtmlContent = await replaceVariableValue(htmlContent, variables, localVariables);
+          const targetHtmlContent = await getRenderContent(engine, htmlContent, variables, localVariables, (data) => {
+            return data;
+          });
           const encodedHtml = encodeURIComponent(targetHtmlContent);
           const dataUrl = 'data:text/html;charset=utf-8,' + encodedHtml;
+
           setSrc(dataUrl);
         } else {
           try {
